==> R Markdown files/0. General Behavior.Rmd <==
---
title: "1. General Behavior"
author: "varholick"
date: "1/3/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#plots
library(tidyverse)
library(ggplot2)
library(ggbeeswarm)
library(ggthemr) #devtools::install_github('Mikata-Project/ggthemr')
library(ggsci)
library(ggpubr)
library(cowplot)
library(readxl)

#stats
library(psych)
library(lme4)
library(lmerTest)
library(MuMIn)

ggthemr('flat')
```

# General incidence of agonistic behavior 
##Extract data from BORIS
1 Open project file in BORIS
2 Click the Observations Menu > Export Events > aggregated events
3 Select the observations you wish to export, then all subjects and all behaviors
4 Click YES for group events into one file
5 Save as .csv file

One file for gg and one for jv*
*select all gg, and then only select the necessary obs for jv

```{r, import data from .csv and organize}
data_master <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/Ag_Boris_20211109_fromMaster_B1B2B3.csv")
not_reliability <- c("jv2", "gg2")

data_1 <- data_master %>% 
  separate(col=Observation.id, into=c("Cage", "Date", "Hour", "Coder"), sep="_") %>% 
  filter(!Coder %in% not_reliability) %>% 
  select(Cage, Date, Hour, Coder, Behavior, Start..s.)

head(data_1)
levels(as.factor(data_1$Coder))

data_2 <- data_1 %>% 
  filter(Cage!="1") %>% 
  filter(Date!=618) %>% 
  filter(Cage!="925878") %>% 
  droplevels()

data_2$Cage <- as.factor(data_2$Cage)
data_2$Date <- as.factor(data_2$Date)
data_2$Hour <- as.factor(data_2$Hour)
data_2$Coder <- as.factor(data_2$Coder)
levels(data_2$Cage)
levels(data_2$Date)
levels(data_2$Hour)
levels(data_2$Coder)
data_2 <- mutate(data_2, minute=round(Start..s./60))

sex_id <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/AcomysDomRegen_Inventory.xlsx", sheet = "Sheet3")
sex_id$Cage <- as.factor(sex_id$Cage)

data_3 <- left_join(data_2, sex_id, by="Cage")
```
## Plot activity across hours
```{r, filter batch1, make graph}
batch1 <- c(957162, 922587, 941702, 927693)
batch1data <- data_3 %>% 
  filter(Cage %in% batch1) %>% 
  filter(Behavior=="Activity") %>% 
  group_by(Cage, Sex, Date, Hour) %>% 
  count()
batch1data <- mutate(batch1data, propact=(n/15)*100)
batch1data <- mutate(batch1data, species="Acomys")
batch1data$Hour <- factor(batch1data$Hour,levels=c("20","21","22","23","00","01","02","03","04","05","06","07","08","09","10","11","12","13","14","15","16","17","18","19"))


activity_plot <- ggplot(batch1data, aes(Hour, propact, group=species, color=Sex, shape=Sex)) +
  #stat_summary(fun=median, geom="line") +
  geom_pointrange(stat="summary",
                  fun.min=function(z) {quantile(z,0.25)},
                  fun.max=function(z) {quantile(z,0.75)},
                  fun=median)+
  geom_vline(xintercept="06", linetype="dashed", color="black")+
  ylab("Perc time active") + xlab("Hour of day")+
  scale_color_npg()+
  #scale_color_manual(values=c("#3498DB","#2ECC71")) +
  scale_x_discrete(labels=c("20","","","","00","","","","04","","","","08","","","","12","","","","16","","",""))+
  facet_grid(rows=vars(Sex))+
  guides(color="none", shape="none")
 
activity_plot
```
## Boxplot of agonistic behaviors during dark cycle
```{r, }
lightness <- c("06","07","08","09","10","11","12","13","14","15","16","17","18","19")
data_night <- data_3 %>% 
  filter(!Hour %in% lightness)

Batch2 <- c("1001363", "990342", "925878")

agbhv <- data_night %>% 
  filter(Behavior!="Activity") %>% 
  group_by(Cage, Sex, Date, Hour, Behavior) %>% 
  count() %>% 
  mutate(prop=(n/15)*100)

agbhv <- filter(agbhv, !Cage %in% Batch2)

agbhv$Behavior <- factor(agbhv$Behavior, levels=c("Chasing","Induced Flee", "Mounting (active)", "Side Huddle", "Mounted Huddle"))

agbhv$n <- as.numeric(agbhv$n)

not_ag <- c("Side Huddle", "Mounted Huddle", "Activity")
nightag <- data_night %>% 
  filter(!Behavior %in% not_ag) %>% 
  mutate(Behavior2="Agonistic") %>% 
  group_by(Cage, Sex, Date, Hour, Behavior2) %>% 
  count() %>% 
  mutate(prop=(n/15)*100)


zeromatrix <- data_night %>% 
  select(Cage, Sex, Date, Hour) %>%
  distinct() %>% 
  mutate(Behavior2="Agonistic") %>% 
  mutate(n=0) %>% 
  filter(Hour=="20")

nightag <- rbind(nightag, zeromatrix)
nightag$Hour <- factor(nightag$Hour,levels=c("20","21","22","23","00","01","02","03","04","05"))

night_plot <- ggplot(nightag, aes(Hour, n, group=Sex, color=Sex, shape=Sex)) +
  geom_pointrange(stat="summary",
                  fun.min=function(z) {quantile(z,0.25)},
                  fun.max=function(z) {quantile(z,0.75)},
                  fun=median, 
                  position=position_dodge(width=0.5))+
  ylab("frequency ag bhv") + xlab("Hour of night")+
  scale_color_npg()+
  labs(color="Sex", shape="Sex")


figure1 <- ggdraw() + 
  draw_plot(activity_plot, x=0, y=0, width=0.4, height=1) +
  draw_plot(night_plot, x=.4, y=0, width=0.6, height=1)+
  draw_plot_label(label=c("A","B"), size=12, x=c(0,0.4), y=c(1,1))

figure1

ggsave("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/R generated figures/SI_activity_plot.tiff", figure1, units="cm", width = 20, height = 10, device='tiff', dpi=300)


```
