==> R Markdown files/5.1 Differences ind-level social dom bhv.Rmd <==
---
title: "5. Differences ind-level social dom bhv"
author: "Justin Varholick"
date: '2022-06-14'
output: html_document
---

```{r setup}
library(tidyverse)
library(readxl)

library(lubridate)
library(igraph)

library(ggplot2)
library(ggsci)
library(cowplot)
library(stringr)
```
##Extract data from BORIS
1 Open project file in BORIS
2 Click the Observations Menu > Export Events > aggregated events
3 Select the observations you wish to export, then all subjects and all behaviors
4 Click YES for group events into one file
5 Save as .csv file
6 edit in excel 
  1. split by "_" delimiter into 6 columns
  2. remove "MVI" from data
  3. Check all dates are actual dates (remove "2020")
  4. check all hours are actual hours (replace pt1, and shift 21 and 12 which extend to next column)
  5. check that all columns have the correct data, delete the coder initials data

```{r import data}
adata <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/dominanceevents_MASTER_adj.csv", stringsAsFactors=TRUE)

#Fixing observation.id inconsistencies...
adata$Cage <- adata$Observation.id
adata$Subject <- gsub("Mouse ","", as.factor(adata$Subject))
adata$Date <- gsub("-Jan","01", adata$Date)
adata$Modifiers <- gsub("None", "3", adata$Modifiers)
adata$Date <- gsub("3001", "20200130", adata$Date)
adata$Date <- gsub("2901", "20200129", adata$Date)
adata$Date <- gsub("2801", "20200128", adata$Date)
adata$Date <- gsub("2301", "20200123", adata$Date)
adata$Date <- gsub("2201", "20200122", adata$Date)
adata$Date <- gsub("2101", "20200121", adata$Date)
adata$Date <- gsub("730", "20190730", adata$Date)
adata$Date <- gsub("731", "20190731", adata$Date)
adata$Date <- gsub("801", "20190801", adata$Date)
adata$Date <- gsub("806", "20190806", adata$Date)
adata$Date <- gsub("807", "20190807", adata$Date)
adata$Date <- gsub("808", "20190808", adata$Date)
adata$Date <- gsub("815", "20190815", adata$Date)
adata$Date <- gsub("816", "20190816", adata$Date)
adata$Date <- gsub("817", "20190817", adata$Date)
adata$Date <- gsub("2019020190731", "20190731", adata$Date)
adata$Date <- gsub("2019020190807", "20190807", adata$Date)
adata$Date <- gsub("916", "20200916", adata$Date)
adata$Date <- gsub("922", "20200922", adata$Date)
adata$Date <- gsub("923", "20200923", adata$Date)
adata$Date <- gsub("2020020200916", "20200916", adata$Date)
adata$Date <- gsub("2020020200922", "20200922", adata$Date)
adata$Date <- gsub("2020020200923", "20200923", adata$Date)

adata$Hour2 <- paste(adata$Hour, "00", sep=":")
adata$Date <- ymd(adata$Date)
adata$DateHour <- paste(adata$Date, adata$Hour2, sep=" ")
adata$DateHour <- as.POSIXct(adata$DateHour, tz="America/New_York")
adata$DateHour <- ymd_hms(adata$DateHour)

adata$Start..s. <- duration(adata$Start..s., "seconds")
adata$Stop..s. <- duration(adata$Stop..s., "seconds")

adata$Start <- (adata$DateHour + adata$Start..s.)
adata$Stop <- (adata$DateHour + adata$Stop..s.)

ag <- adata %>% 
  within(Opponent.1 <- paste(Cage, Subject, sep="-")) %>% 
  within(value <- paste("1")) %>% 
  select(Opponent.1, value, Date, Behavior, Behavioral.category, Modifiers, Cage, Week)

locations <- ag %>% 
  filter(Behavioral.category=="Location") %>% 
  filter(Behavior!="Shelter") %>% 
  filter(Behavior!="Food Hopper") %>% 
  filter(Behavior!="Cup") %>% 
  unite("Behavior2", Behavior, Modifiers, sep="-", remove=T) %>% 
  rename("Behavior"=Behavior2)

ag <- ag %>% 
  filter(Behavioral.category!="Location") %>% 
  select(Opponent.1, value, Date, Behavior, Behavioral.category, Cage, Week)
ag <- rbind(ag, locations)

#add independent variables
sex_id <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/AcomysDomRegen_Inventory.xlsx", sheet = "Sheet3")
sex_id$Cage <- as.factor(sex_id$Cage)
ag$Cage <- as.factor(ag$Cage)

ag <- left_join(ag, sex_id, by="Cage")

ind_ds <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/dscorebyanimal_edited.csv")
ind_ds$Cage <- as.factor(ind_ds$Cage)
ind_ds <- rename(ind_ds, "Opponent.1"=AnimalID)

ag <- left_join(ag, ind_ds, by=c("Opponent.1", "Cage"))
```


```{r plot, compare dyads}
#data organization
ag2 <- ag %>% 
  group_by(Cage, Opponent.1, Sex, GrpSize, Age, dom_status_ds, Behavior) %>% 
  count(value)

zeromatrix <- ag %>%
  select(Opponent.1, Sex, GrpSize, Age, Cage, dom_status_ds) %>% 
  unique() %>% 
  mutate(Chasing=0) %>% 
  mutate(Mounting=0) %>% 
  mutate(Attacking=0) %>% 
  mutate(Stealing=0) %>% 
  mutate(Fleeing=0) %>% 
  mutate(`Induced Flee`=0) %>%
  mutate(Freeze=0) %>% 
  mutate(Unseen=0) %>% 
  mutate(Inactive=0) %>% 
  mutate(`Top-Left`=0) %>% 
  mutate(`Top-Right`=0) %>% 
  mutate(`Bottom-Left`=0) %>% 
  mutate(`Bottom-Right`=0) %>% 
  gather("Behavior", "n_prime", 7:19)

ag_zeros <- left_join(zeromatrix, ag2, by=(c("Opponent.1", "Sex", "GrpSize", "Age", "dom_status_ds", "Cage", "Behavior")))
ag_zeros[is.na(ag_zeros)]<-0

ag3 <- ag_zeros %>% 
  filter(Behavior!="Top-Right") %>% 
  filter(Behavior!="Top-Left") %>% 
  filter(Behavior!="Bottom-Right") %>% 
  filter(Behavior!="Bottom-Left")

ag4 <- ag3 %>% 
  filter(Behavior=="Induced Flee") %>% 
  mutate(Behavior="Induced \nFlee")

ag5 <- ag3 %>% 
  filter(Behavior!="Induced Flee")

ag3 <- rbind(ag4, ag5)

ag3$Behavior <- factor(ag3$Behavior, levels=c("Chasing", "Mounting", "Attacking", "Stealing", "Fleeing", "Induced \nFlee", "Freeze", "Unseen", "Inactive"))
```


```{r dom off vs sub def}
off <- c("Chasing", "Attacking", "Mounting", "Stealing")
def <- c("Fleeing", "Induced \nFlee", "Freeze")

ag_dom <- ag3 %>% 
  filter(Behavior %in% off) %>% 
  filter(dom_status_ds=="Dominant") %>% 
  group_by(Cage, Sex, GrpSize, Age) %>% 
  summarise(dom_off_n=sum(n))

ag_sub <- ag3 %>% 
  filter(Behavior %in% def) %>% 
  filter(dom_status_ds=="Subordinate") %>% 
  group_by(Cage, Sex, GrpSize, Age) %>% 
  summarise(sub_def_n=sum(n))

ag_domsub <- left_join(ag_dom, ag_sub, by=c("Cage", "Sex", "GrpSize", "Age"))

ag_domsub2 <- gather(ag_domsub, "group", "n", 5:6)
```
#Compare dyads
```{r}
sex_dyads_dom <- ag3 %>% 
  filter(GrpSize==2) %>% 
  filter(Age=="Young") %>% 
  filter(dom_status_ds=="Dominant") %>% 
  filter(Behavior %in% off)

ggplot(sex_dyads_dom)+
  stat_summary(mapping=aes(Behavior, n, fill=Sex), geom="col", fun="mean", color="black", position=position_dodge2(width=0.90))+
  geom_point(mapping=aes(Behavior, n, group=Sex), color="black", position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, n, group=Sex), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_fill_npg()
```
#Test, sex
```{r}
m.sex.dom <- lmer(n ~ (1|Opponent.1) + Behavior + Sex + Behavior*Sex, data=sex_dyads_dom)
anova(m.sex.dom, test="F")
summary(m.sex.dom)
```
#Sex, sub
```{r}
sex_dyads_sub <- ag3 %>% 
  filter(GrpSize==2) %>% 
  filter(Age=="Young") %>% 
  filter(dom_status_ds=="Subordinate") %>% 
  filter(Behavior %in% def)

#need to set freezing as the reference group for the stat model
sex_dyads_sub$Behavior <- factor(sex_dyads_sub$Behavior, levels=c("Freeze", "Induced \nFlee", "Fleeing", "Chasing", "Mounting", "Attacking", "(Other)"))

ggplot(sex_dyads_sub, aes(Behavior, n, fill=Sex))+
  geom_boxplot()+
  scale_fill_npg()

ggplot(sex_dyads_sub)+
  stat_summary(mapping=aes(Behavior, n, fill=Sex), geom="col", fun="mean", color="black", position=position_dodge2(width=0.90))+
  geom_point(mapping=aes(Behavior, n, group=Sex), color="black", position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, n, group=Sex), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_fill_npg()
```
#Test, sex sub
```{r}
m.sex.sub <- lmer(n ~ (1|Opponent.1) + Behavior + Sex + Behavior*Sex, data=sex_dyads_sub)
anova(m.sex.sub)
summary(m.sex.sub)
```



#Sex, dom sub
```{r}
sex_dyads_domsub <- ag_domsub2 %>% 
  filter(GrpSize==2) %>% 
  filter(Age=="Young")

ggplot(sex_dyads_domsub, aes(group, n, fill=Sex, color=Sex, group=Cage)) +
  geom_point() + geom_line()
```
#test, sex domsub
```{r}
m.sex.domsub <- lmer(n ~ (1|Cage) + group + Sex + group*Sex, data=sex_dyads_domsub)
anova(m.sex.domsub, test="F")
summary(m.sex.domsub)


```


##Group Size Effects, Males
```{r}
males_dom <- ag3 %>% 
  filter(Sex=="M") %>% 
  filter(dom_status_ds=="Dominant") %>% 
  filter(Behavior %in% off)

ggplot(males_dom, aes(Behavior, n, fill=as.factor(GrpSize)))+
  geom_boxplot()+
  scale_fill_npg()
```
#Test, grpsize
```{r}
m.grp.dom <- lmer(n ~ (1|Opponent.1) + Behavior + as.factor(GrpSize) + Behavior*as.factor(GrpSize), data=males_dom)
anova(m.grp.dom, test="F")
summary(m.grp.dom)
```
#Grpsize, sub
```{r}
males_sub <- ag3 %>% 
  filter(Sex=="M") %>% 
  filter(dom_status_ds=="Subordinate") %>% 
  filter(Behavior %in% def)

#need to set freezing as the reference group for the stat model
males_sub$Behavior <- factor(males_sub$Behavior, levels=c("Freeze", "Induced \nFlee", "Fleeing", "Chasing", "Mounting", "Attacking", "(Other)"))

ggplot(males_sub, aes(Behavior, n, fill=as.factor(GrpSize)))+
  geom_boxplot()+
  scale_fill_npg()
```
#Test, grpsize sub
```{r}
m.grp.sub <- lmer(n ~ (1|Opponent.1) + Behavior + as.factor(GrpSize) + Behavior*as.factor(GrpSize), data=males_sub)
anova(m.grp.sub, test="F")
summary(m.grp.sub)
```
```{r}
males_domsub <- ag_domsub2 %>% 
  filter(Sex=="M")

ggplot(males_domsub, aes(group, n, fill=as.factor(GrpSize), color=as.factor(GrpSize), group=Cage)) +
  geom_point() + geom_line()
```
```{r}
m.males.domsub <- lmer(n ~ (1|Cage) + group + as.factor(GrpSize) + group*as.factor(GrpSize), data=males_domsub)
anova(m.males.domsub, test="F")
summary(m.males.domsub)
```



```{r grouped plot}
dom_sex_plot <- ggplot(sex_dyads_dom)+
  stat_summary(mapping=aes(Behavior, n, fill=Sex), geom="col", fun="mean", color="black", position=position_dodge2(width=0.90))+
  #geom_point(mapping=aes(Behavior, n, group=Sex), color="black", size=0.5, position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, n, group=Sex), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_y_continuous(limits=c(0,200), expand=c(0,0))+
  scale_fill_npg()+
  ylab("Total freq.") + xlab("")+
  theme_cowplot()+
  theme(legend.position = "none")

sub_sex_plot <- ggplot(sex_dyads_sub)+
  stat_summary(mapping=aes(Behavior, n, fill=Sex), geom="col", fun="mean", color="black", position=position_dodge2(width=0.90))+
  #geom_point(mapping=aes(Behavior, n, group=Sex), color="black", size=0.5, position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, n, group=Sex), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_y_continuous(limits=c(0,200), expand=c(0,0))+
  ylab("Total freq.") + xlab("")+
  scale_fill_npg()+
  theme_cowplot()

males_plot <- ggplot(males_dom, aes(Behavior, n, color=as.factor(GrpSize), fill=as.factor(GrpSize)))+
  stat_summary(mapping=aes(Behavior, n, fill=as.factor(GrpSize)), geom="col", fun="mean", color="black", position=position_dodge2(width=0.90))+
  #geom_point(mapping=aes(Behavior, n, group=as.factor(GrpSize)), color="black", size=0.5, position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, n, group=as.factor(GrpSize)), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_y_continuous(limits=c(0,200), expand=c(0,0))+
  scale_color_manual(values=c("#00A087FF", "#3C5488FF"))+
  scale_fill_manual(values=c("#00A087FF", "#3C5488FF"))+
  ylab("Total freq.") + xlab("")+
  theme_cowplot()+
  theme(legend.position = "none")

males_sub_plot <- ggplot(males_sub)+
  stat_summary(mapping=aes(Behavior, n, fill=as.factor(GrpSize)), geom="col", fun="mean", color="black", position=position_dodge2(width=0.90))+
  #geom_point(mapping=aes(Behavior, n, group=as.factor(GrpSize)), color="black", size=0.5, position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, n, group=as.factor(GrpSize)), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_y_continuous(limits=c(0,200), expand=c(0,0))+
  scale_fill_manual(values=c("#00A087FF", "#3C5488FF"))+
  ylab("Total freq.") + xlab("")+
  guides(fill=guide_legend(str_wrap("Group Size Males", 5)))+
  theme(legend.position = "right")+
  theme_cowplot()


fig3 <- plot_grid(
  dom_sex_plot, sub_sex_plot,
  males_plot, males_sub_plot,
  labels = "AUTO", ncol = 2, align = c("h","v")
)

ggsave("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/R generated figures/Fig3_socbhv_plots.png", fig3, units="in", width = 9, height = 6, device='png', dpi=300)

fig3
```
