==> R Markdown files/6. Differences grp-level ag hdl bhv.Rmd <==
---
title: "2. Differences ind-level social dom bhv"
author: "Justin Varholick"
date: '2022-06-14'
output: html_document
---

```{r setup}
library(tidyverse)
library(readxl)
library(hms)

library(ggplot2)
library(ggsci)
library(cowplot)

library(lme4)
library(lmerTest)
#custom functions
small_wilcox <- function(group1, group2, grouping_variable, prediction, data){
  library(dplyr)
  data_temp <- dplyr::mutate(data, ranks=rank({{ prediction }}))
  data_gp1 <- dplyr::filter(data_temp, {{ grouping_variable }}=={{ group1 }})
  data_gp2 <- dplyr::filter(data_temp, {{ grouping_variable }}=={{ group2 }})
  group1sum <- sum(data_gp1$ranks)
  group2sum <- sum(data_gp2$ranks)
  total_n <- dplyr::count(data, {{ grouping_variable }})
  
  return(c(group1sum, group2sum, total_n))
}

```
##Extract data from BORIS

1.  Open project file in BORIS

2.  Click the Observations Menu \> Export Events \> aggregated events

3.  Select the observations you wish to export, then all subjects and
    all behaviors

4.  Click YES for group events into one file

5.  Save as .csv file

One file for gg and one for jvselect all gg, and then only select the
necessary obs for jv

```{r data import}
#import data
data_master <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/Ag_Boris_20211109_fromMaster_B1B2B3.csv")
not_reliability <- c("jv2", "gg2")

#organize
data_1 <- data_master %>% 
  separate(col=Observation.id, into=c("Cage", "Date", "Hour", "Coder"), sep="_") %>% 
  filter(!Coder %in% not_reliability) %>% 
  select(Cage, Date, Hour, Coder, Behavior, Start..s.)

#hours
data_1$Hour_chr <- data_1$Hour
data_1$Hour <- sub("^0", "", data_1$Hour)
data_1$Hour <- as.numeric(data_1$Hour)
data_1$Hour <- hms::hms(hour=data_1$Hour)

head(data_1)
levels(as.factor(data_1$Coder))
data_1$Cage <- as.factor(data_1$Cage)
data_1$Date <- as.factor(data_1$Date)
data_1$Coder <- as.factor(data_1$Coder)
levels(data_1$Cage)
levels(data_1$Date)
levels(data_1$Coder)
data_1 <- mutate(data_1, minute=round(Start..s./60))

#add sex ID
sex_id <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/AcomysDomRegen_Inventory.xlsx", sheet = "Sheet3")
sex_id$Cage <- as.factor(sex_id$Cage)

data_3 <- left_join(data_1, sex_id, by="Cage")

#add inactivity 
activity <- data_3[,c(1,2,3,8)]
activity <- unique(activity)
activity$Minute <- c(levels(data_1$minute))

#limit to night
lightness <- c(hms(hour=6:19))
data_night <- data_3 %>% 
  filter(!Hour %in% lightness) 

data_night1 <- data_night %>% 
  group_by(Cage, Sex, GrpSize, Age, Date, Hour, Hour_chr, Behavior) %>%
  count() %>% 
  mutate(n=ifelse(n>15,15,n)) %>% #spotted a few cases of 16 and 17
  mutate(prop=(n/15)*100)

zeromatrix <- data_night %>% 
  select(Cage, Sex, GrpSize, Age, Date, Hour, Hour_chr) %>% 
  distinct() %>% 
  mutate(Activity=0) %>%
  mutate(Chasing=0) %>% 
  mutate(`Induced Flee`=0) %>% 
  mutate(`Mounting (active)`=0) %>%
  mutate(`Mounted Huddle`=0) %>% 
  mutate(`Side Huddle`=0) %>% 
  gather("Behavior", "n_prime", 8:13)
  

nightag <- left_join(zeromatrix, data_night1, by=c("Cage", "Sex", "GrpSize", "Age", "Date", "Hour", "Hour_chr", "Behavior")) 
nightag[is.na(nightag)]<-0 
nightag <- nightag[,c(1:8,10:11)] 

nightag$Hour_chr <- factor(nightag$Hour_chr, levels=c("20","21","22","23","00","01","02","03","04","05"))


```


```{r data import}
#add day
pm1 <- c(hms::hms(hour=20:23))
am1 <- c(hms::hms(hour=0:19))

nightag_1 <- nightag %>% 
  mutate("Day"=case_when(Hour %in% pm1 & Date=="20200121" ~ "Day1",
                         Hour %in% am1 & Date=="20200122" ~ "Day1",
                         Hour %in% pm1 & Date=="20200122" ~ "Day2",
                         Hour %in% am1 & Date=="20200123" ~ "Day2",
                         Hour %in% pm1 & Date=="20200123" ~ "Day3",
                         Hour %in% am1 & Date=="20200124" ~ "Day3",
                         
                         Hour %in% pm1 & Date=="20200130" ~ "Day8",
                         Hour %in% am1 & Date=="20200131" ~ "Day8",
                         Hour %in% pm1 & Date=="20200131" ~ "Day9",
                         Hour %in% am1 & Date=="20200201" ~ "Day9",
                         Hour %in% pm1 & Date=="20200201" ~ "Day10",
                         Hour %in% am1 & Date=="20200202" ~ "Day10",
                         Hour %in% pm1 & Date=="20200202" ~ "Day11",
                         Hour %in% am1 & Date=="20200203" ~ "Day11",
                         
                         Hour %in% pm1 & Date=="20200204" ~ "Day12",
                         Hour %in% am1 & Date=="20200205" ~ "Day12",
                         Hour %in% pm1 & Date=="20200205" ~ "Day13",
                         Hour %in% am1 & Date=="20200206" ~ "Day13",
                         Hour %in% pm1 & Date=="20200206" ~ "Day14",
                         Hour %in% am1 & Date=="20200207" ~ "Day14",
                         
                         Hour %in% pm1 & Date=="20190730" ~ "Day1",
                         Hour %in% am1 & Date=="20190731" ~ "Day1",
                         Hour %in% pm1 & Date=="20190731" ~ "Day2",
                         Hour %in% am1 & Date=="20190801" ~ "Day2",
                         Hour %in% pm1 & Date=="20190801" ~ "Day3",
                         Hour %in% am1 & Date=="20190802" ~ "Day3",
                         
                         Hour %in% pm1 & Date=="20190806" ~ "Day8",
                         Hour %in% am1 & Date=="20190807" ~ "Day8",
                         Hour %in% pm1 & Date=="20190807" ~ "Day9",
                         Hour %in% am1 & Date=="20190808" ~ "Day9",
                         Hour %in% pm1 & Date=="20190808" ~ "Day10",
                         Hour %in% am1 & Date=="20190809" ~ "Day10",
                         
                         Hour %in% pm1 & Date=="20190815" ~ "Day15",
                         Hour %in% am1 & Date=="20190816" ~ "Day15",
                         Hour %in% pm1 & Date=="20190816" ~ "Day16",
                         Hour %in% am1 & Date=="20190817" ~ "Day16",
                         Hour %in% pm1 & Date=="20190817" ~ "Day17",
                         Hour %in% am1 & Date=="20190818" ~ "Day17",
                         Hour %in% pm1 & Date=="20190818" ~ "Day18",
                         Hour %in% am1 & Date=="20190819" ~ "Day18",
                         
                         Hour %in% pm1 & Date=="20200915" ~ "Day1",
                         Hour %in% am1 & Date=="20200916" ~ "Day1",
                         Hour %in% pm1 & Date=="20200916" ~ "Day2",
                         Hour %in% am1 & Date=="20200917" ~ "Day2",
                         Hour %in% pm1 & Date=="20200917" ~ "Day3",
                         Hour %in% am1 & Date=="20200918" ~ "Day3",
                         
                         Hour %in% pm1 & Date=="20200922" ~ "Day8",
                         Hour %in% am1 & Date=="20200923" ~ "Day8",
                         Hour %in% pm1 & Date=="20200923" ~ "Day9",
                         Hour %in% am1 & Date=="20200924" ~ "Day9",
                         Hour %in% pm1 & Date=="20200924" ~ "Day10",
                         Hour %in% am1 & Date=="20200925" ~ "Day10",
                         
                         Hour %in% pm1 & Date=="20200929" ~ "Day15",
                         Hour %in% am1 & Date=="20200930" ~ "Day15",
                         Hour %in% pm1 & Date=="20200930" ~ "Day16",
                         Hour %in% am1 & Date=="20201001" ~ "Day16",
                         Hour %in% pm1 & Date=="20201001" ~ "Day17",
                         Hour %in% am1 & Date=="20201002" ~ "Day17"))

#remove weird observation times, why man, why?
crazydays <- c("Day11", "Day12", "Day13", "Day14", "Day18")

nightag_1 <- filter(nightag_1, !Day %in% crazydays)

```

```{r}
grouped_data_day <- nightag_1 %>% 
  group_by(Cage, Sex, GrpSize, Age, Day, Behavior) %>% 
  summarise(total_minutes_day=sum(n)) %>% 
  mutate(prop_day=(total_minutes_day/150))
grouped_data_cage <- nightag_1 %>% 
  group_by(Cage, Sex, GrpSize, Age, Behavior) %>% 
  summarise(total_minutes_cage=sum(n)) %>% 
  mutate(prop_cage=(total_minutes_cage/1350))


```

##Dyads
```{r}
sex_dyads<- grouped_data_cage %>% 
  filter(GrpSize==2) %>% 
  filter(Age=="Young") %>% 
  filter(Behavior!="Mounting (active)") %>% 
  filter(Behavior!="Activity")

ggplot(sex_dyads, aes(Behavior, prop_cage, fill=Sex))+
  geom_boxplot()+
  scale_y_continuous(limits = c(0,1))+
  scale_fill_npg()

```
#test sex
```{r}
m.sex <- glm(prop_cage ~ Behavior + Sex + Behavior*Sex, data=sex_dyads)
anova(m.sex, test="F")
summary(m.sex)
```

##Males
```{r}
males<- grouped_data_cage %>% 
  filter(Sex=="M") %>% 
  filter(Behavior!="Mounting (active)") %>% 
  filter(Behavior!="Activity")

ggplot(males, aes(Behavior, prop_cage, fill=as.factor(GrpSize)))+
  geom_boxplot()+
  scale_y_continuous(limits = c(0,1))+
  scale_fill_npg()

```
#Test grp size
```{r}
m.grp <- glm(prop_cage ~ Behavior + as.factor(GrpSize) + Behavior*as.factor(GrpSize), data=males)
anova(m.grp, test="F")
summary(m.grp)
```

##Age
```{r}
females<- grouped_data_cage %>% 
  filter(Sex=="F") %>% 
  filter(Behavior!="Mounting (active)") %>% 
  filter(Behavior!="Activity")

ggplot(females, aes(Behavior, prop_cage, fill=Age))+
  geom_boxplot()+
  scale_y_continuous(limits = c(0,1))+
  scale_fill_npg()

```
#Test age
```{r}
m.age <- glm(prop_cage ~ Behavior + Age + Behavior*Age, data=females)
anova(m.age, test="F") 
summary(m.age)
```

#plots
```{r}
grp_plot <- ggplot(males)+
  stat_summary(mapping=aes(Behavior, prop_cage, fill=as.factor(GrpSize)), geom="col", fun="mean", position=position_dodge2(width=0.90))+
  #geom_point(mapping=aes(Behavior, prop_cage, group=as.factor(GrpSize)), size=0.5, color="black", position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, prop_cage, group=as.factor(GrpSize)), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_y_continuous(limits = c(0,1), expand=c(0,0))+
  scale_color_manual(values=c("#00A087FF", "#3C5488FF"))+
  scale_fill_manual(values=c("#00A087FF", "#3C5488FF"))+
  ylab("proportion of \nbehavior per cage") + xlab("")+
  guides(fill=guide_legend(str_wrap("Group Size Males", 5)), color=guide_legend(str_wrap("Group Size", 5))) + theme_cowplot()


age_plot <- ggplot(females)+
  stat_summary(mapping=aes(Behavior, prop_cage, fill=Age), geom="col", fun="mean", position=position_dodge2(width=0.90))+
  #geom_point(mapping=aes(Behavior, prop_cage, group=Age), size=0.5, color="black", position=position_dodge2(width=0.90))+
  stat_summary(mapping=aes(Behavior, prop_cage, group=Age), geom="errorbar", fun.data="mean_se", color="black", size=0.5, position=position_dodge2(width=0.90))+
  scale_y_continuous(limits = c(0,1), expand=c(0,0))+
  scale_color_manual(values=c("#F39B7FFF", "#8491B4FF"))+
  scale_fill_manual(values=c("#F39B7FFF", "#8491B4FF"))+
  ylab("proportion of \nbehavior per cage") + xlab("")+
  guides(fill=guide_legend(str_wrap("Age Females", 5)), color=guide_legend(str_wrap("Age Females", 5)))+theme_cowplot()

fig4 <- plot_grid(
  grp_plot, age_plot,
  labels="AUTO", ncol=1, align="v"
)

ggsave("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/R generated figures/Fig4_nightbhv_plots.png", fig4, units="in", width = 9, height = 6, device='png', dpi=300)

fig4
```
