==> R Markdown files/7. Ear-hole regeneration.Rmd <==
---
title: "7. Ear-hole regeneration"
author: "varholick"
date: "8/3/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#plots
library(tidyverse)
library(ggplot2)
library(ggbeeswarm)
library(ggsci)
library(ggpubr)
library(cowplot)
library(readxl)
library(scales)

#stats
library(psych)
library(lme4)
library(lmerTest)
library(MuMIn)
```
#organize data
```{r}
eardata <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/EarDaysToClose.xlsx")
sex_id <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/AcomysDomRegen_Inventory.xlsx", sheet = "Sheet1")
ind_ds <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/dscorebyanimal_edited.csv")
stability <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/stabilitybycage.csv")

eardata <- left_join(eardata, sex_id[,1:16], by="AnimalID")
eardata <- left_join(eardata, ind_ds, by="AnimalID")
eardata <- left_join(eardata, stability, by="AnimalID")

eardata$Sex <- as.factor(eardata$Sex)
eardata$Cage.x <- as.factor(eardata$Cage.x)
eardata$GrpSize <- as.factor(eardata$GrpSize)
eardata$dom_status_ds <- as.factor(eardata$dom_status_ds)
eardata$stab_status2 <- factor(eardata$stab_status2, levels=c("Enforced stable", "Unstable", "Unenforced stable", "Unmeasurable"))
```

#mixed model
```{r}
#Models
m.earclose.ageonly <- lmer(DaysToClose ~ (1|Cage.x) + AgeAtTest, data=eardata)
m.earclose.full <- lmer(DaysToClose ~ (1|Cage.x) + AgeAtTest + ds + stab_status2 + ds*stab_status2, data=eardata)

print("Age only Model")
anova(m.earclose.ageonly)
summary(m.earclose.ageonly)
print("-----------------------------------------------------------------------")
print("Full model")
anova(m.earclose.full)
summary(m.earclose.full)
print("-----------------------------------------------------------------------")
print("variance only age")
r.squaredGLMM(m.earclose.ageonly)
print("variance full model")
r.squaredGLMM(m.earclose.full)
```
```{r}
#colinEARity 
st2a.m <- lmer(AgeAtTest ~ stab_status2 + (1|Cage.x), eardata)
print("Stability on Age")
anova(st2a.m)
summary(st2a.m)
print("-----------------------------------------------------------------------")
car::vif(m.earclose.full)
print("-----------------------------------------------------------------------")
dropst.m <- lmer(DaysToClose ~ AgeAtTest + ds + (1|Cage.x), eardata)
anova(dropst.m)
r.squaredGLMM(dropst.m)
car::vif(dropst.m)
```
```{r}
ds_ear_plot <- ggplot(eardata, aes(ds, DaysToClose))+
  geom_point(aes(color=ds), size=2)+
  geom_point(color="black", shape=1, size=2)+
  geom_smooth(color="black", method=lm, se=FALSE, size=1)+
  scale_x_continuous(limits=c(-3,3), breaks=c(-3,-2,-1,0,1,2,3))+
  scale_y_continuous(limits=c(0,115), breaks=seq(0,110, by=10), expand=c(0,0))+
  scale_color_gsea()+
  xlab("David's Score") + ylab("Days to close ear")+
  #annotate("text", x = c(-2,2), y = c(2,2), label = c("more subordinate","more dominant"))+
  labs(color="David's \nScore")+
  theme_cowplot()
ds_ear_plot
```
```{r}
eardata$stab_status2 <- factor(eardata$stab_status2, levels=c("Enforced stable", "Unstable", "Unenforced stable", "Unmeasurable"))

stbl_ear_plot <- ggplot(eardata)+
  stat_summary(mapping=aes(stab_status2, DaysToClose, fill=stab_status2), geom="col", fun="mean", color="black")+
  stat_summary(mapping=aes(stab_status2, DaysToClose), geom="errorbar", fun.data="mean_se",width=0.5)+
  scale_fill_npg()+
  scale_y_continuous(limits=c(0,115), breaks=seq(0,110, by=10), expand=c(0,0))+
  scale_x_discrete(labels=wrap_format(10))+
  xlab("") + ylab("Days to close ear")+
  guides(fill=guide_legend(title="Stability \nCategory", nrow=2, byrow=TRUE))+
  theme_cowplot()+
  theme(legend.position="bottom") 

stbl_ear_plot


```


```{r}
earplot <- ggplot(eardata, aes(ds, DaysToClose))+
  geom_point(aes(color=stab_status2), size=2)+
  geom_point(color="black", shape=1, size=2)+
  geom_smooth(aes(color=stab_status2), method=lm, se=FALSE, size=1)+
  scale_x_continuous(limits=c(-3,3), breaks=c(-3,-2,-1,0,1,2,3))+
  scale_y_continuous(limits=c(0,115), breaks=seq(0,110, by=10), expand=c(0,0))+
  scale_color_npg()+
  xlab("David's \nScore") + ylab("Days to close ear")+
  #annotate("text", x = c(-2,2), y = c(2,2), label = c("more subordinate","more dominant"))+
  labs(color="Stability Category", shape="Stability Category")+
  theme_cowplot()+
  theme(legend.position="none") 
earplot
```

```{r, age plot}
ageplot <- ggplot(eardata, aes(AgeAtTest, DaysToClose))+
  geom_point(aes(color=stab_status2), size=4)+
  scale_color_npg()+
  xlab("Age at experiment d0") + ylab("Days to close ear")+
  labs(color="Stability Category")
ageplot
```

```{r}
ageplot2 <- ggplot(eardata, aes(AgeAtTest, DaysToClose))+
  scale_y_continuous(limits=c(0,115), breaks=seq(0,110, by=10), expand=c(0,0))+
  geom_smooth(method="lm", color="black", se=F)+
  geom_point(size=2, color="black")+
  scale_color_npg()+
  xlab("Age in days") + ylab("Days to close ear")+
  theme_cowplot()
ageplot2
```
```{r}
ear_plot <- plot_grid(
  ageplot2, 
  ds_ear_plot+ylab("")+theme(legend.position = "none", axis.text.y=element_blank()),
                             stbl_ear_plot + theme(legend.position = "none"), 
                             earplot + ylab("")+theme(axis.text.y=element_blank()),
                                                                       ncol=2, labels="AUTO")
 
ggsave("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/R generated figures/fig5.1_ear.png", ear_plot, units="in", width = 6, height = 6, device='png', dpi=300)

ear_plot
```


```{r}
ggsave("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/R generated figures/SIFig_age_ear.tiff", ageplot, units="in", width = 6, height = 3, device='tiff', dpi=300)
```
