==> R Markdown files/8. blood serum.Rmd <==
---
title: "8. blood serum"
author: "varholick"
date: "8/3/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#plots
library(tidyverse)
library(ggplot2)
library(ggbeeswarm)
library(ggsci)
library(ggpubr)
library(cowplot)
library(readxl)

#stats
library(lme4)
library(lmerTest)
library(MuMIn)

```

# data organization
```{r}
cort_data <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/cort_combined.csv", header=TRUE)

sex_id <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/AcomysDomRegen_Inventory.xlsx", sheet = "Sheet1")
ind_ds <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/dscorebyanimal_edited.csv")
eardata <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/EarDaysToClose.xlsx")
stability <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/stabilitybycage.csv")

cort_data <- left_join(cort_data, sex_id[,1:16], by="AnimalID")
cort_data <- left_join(cort_data, ind_ds, by="AnimalID")
cort_data <- left_join(cort_data, stability, by="AnimalID")
cort_data <- left_join(cort_data, eardata, by="AnimalID")

cort_data$Sex.x <- as.factor(cort_data$Sex.x)
cort_data$Cage.x <- as.factor(cort_data$Cage.x)
cort_data$GrpSize <- as.factor(cort_data$GrpSize)
cort_data$dom_status_ds <- as.factor(cort_data$dom_status_ds)
cort_data$stab_status <- factor(cort_data$stab_status, levels=c("Enforced stable", "Unstable", "Unenforced stable", "Unmeasurable"))

cort_data <- filter(cort_data, Sex.x!="F")
```

```{r}
m.cort.full <- lmer(ng ~ (1|Cage.x) + ds + stab_status + ds*stab_status, data=cort_data)
m.cort.simp <- lm(ng ~ ds + stab_status + ds*stab_status, data=cort_data)


anova(m.cort.full)
anova(m.cort.simp)

r.squaredGLMM(m.cort.simp)
r.squaredGLMM(m.cort.full)
```

```{r}
cor.test(cort_data$DaysToClose, cort_data$ng, method="spearman", exact=FALSE)
```
```{r}
age_blood_plot <- ggplot(cort_data, aes(AgeAtTest, ng))+
  geom_smooth(method="lm", color="black", fill="darkgrey")+
  geom_jitter(size=2, color="black")+
  stat_cor(method="spearman", alternative="two.sided", p.accuracy = 0.001,
           label.x=100, label.y=900)+
  scale_y_continuous(limits=c(0,1020), expand=c(0,0))+
  scale_color_npg()+
  xlab("Age in days") + ylab("ng/ml of Cortisol")+
  theme(legend.position="right")+theme_cowplot()
age_blood_plot
```
```{r}
ds_blood_plot <- ggplot(cort_data, aes(ds, ng))+
  geom_point(aes(color=ds), size=2)+
  geom_point(color="black", shape=1, size=2)+
  geom_smooth(color="black", method=lm, se=FALSE, size=1)+
  scale_x_continuous(limits=c(-3,3), breaks=c(-3,-2,-1,0,1,2,3))+
  scale_y_continuous(limits=c(0,1020), expand=c(0,0))+
  scale_color_gsea()+
  xlab("David's \nScore") + ylab("ng/ml of Cortisol")+
 # annotate("text", x = c(-2,2), y = c(2,2), label = c("more subordinate","more dominant"))+
  labs(color="David's \nScore")+
  theme_cowplot()
ds_blood_plot
```
```{r}
cort_data$stab_status2 <- factor(cort_data$stab_status2, levels=c("Enforced stable", "Unstable", "Unenforced stable", "Unmeasurable"))

stbl_blood_plot <- ggplot(cort_data)+
  stat_summary(mapping=aes(stab_status2, ng, fill=stab_status2), geom="col", fun="mean", color="black")+
  stat_summary(mapping=aes(stab_status2, ng), geom="errorbar", fun.data="mean_se",width=0.5)+
  scale_fill_npg()+
  xlab("") + ylab("ng/ml of Cortisol")+
  scale_y_continuous(limits=c(0,1020), expand=c(0,0))+
  scale_x_discrete(labels=wrap_format(10))+
  guides(fill=guide_legend(title="Stability \nCategory", nrow=2, byrow=TRUE))+
  theme(legend.position="bottom") +
  theme_cowplot()

stbl_blood_plot
```


```{r}
cortplot <- ggplot(cort_data, aes(ds, ng))+
  geom_point(aes(color=stab_status), size=2)+
  geom_point(color="black", shape=1, size=2)+
  geom_smooth(aes(color=stab_status), method=lm, se=FALSE, size=1)+
  scale_x_continuous(limits=c(-3,3), breaks=c(-3,-2,-1,0,1,2,3))+
  scale_y_continuous(limits=c(0,1020), expand=c(0,0))+
  scale_color_npg()+
  xlab("David's Score") + ylab("ng of Cortisol")+
  #annotate("text", x = c(-2.485,2.525), y = c(2,2), label = c("more subordinate","more dominant"))+
  labs(color="Stability Category")+
  theme_cowplot()
cortplot
```

```{r}
corrplot<- ggplot(cort_data, aes(DaysToClose, ng)) +
  geom_jitter(size=2, color="black") +
  geom_smooth(method="lm", color="black", fill="darkgrey")+
  stat_cor(method="spearman", alternative="two.sided", p.accuracy = 0.001,
           label.x=40, label.y=900)+
  scale_y_continuous(limits=c(0,1020), expand=c(0,0))+
  xlab("Days to close ear") + ylab("ng of Cortisol")+
  theme_cowplot()
corrplot
```
```{r}
fig6 <- plot_grid(
  ds_blood_plot+theme(legend.position="none"), 
  stbl_blood_plot+ylab("")+theme(legend.position="none", axis.text.y=element_blank()),
  cortplot+theme(legend.position="none"), 
  age_blood_plot+ylab("")+theme(axis.text.y=element_blank()), 
  corrplot,
  labels="AUTO", ncol=2
)

ggsave("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/R generated figures/Fig6_cort_plots.png", fig6, units="in", width = 6, height = 9, device='png', dpi=300)

fig6
```
