==> R Markdown files/10. qPCR (cont.).Rmd <==
---
title: "10. qPCR (cont.)"
author: "varholick"
date: "1/3/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#plots
library(tidyverse)
library(ggplot2)
library(ggbeeswarm)
library(ggsci)
library(ggpubr)
library(cowplot)
library(readxl)
library(ggsignif)

#stats
library(lme4)
library(lmerTest)
library(MuMIn)

```
```{r}
qpcr_data <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/qpcr_combined_041123.csv")
sex_id <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/AcomysDomRegen_Inventory.xlsx", sheet = "Sheet1")
ind_ds <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/dscorebyanimal_edited.csv")
eardata <- read_excel("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/EarDaysToClose.xlsx")
stability <- read.csv("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/Data/stabilitybycage.csv")


qpcr_data <- left_join(qpcr_data, sex_id[,c(4,6,15)], by="AnimalID")
qpcr_data <- left_join(qpcr_data, ind_ds[,2:5], by="AnimalID")
qpcr_data <- left_join(qpcr_data, eardata, by="AnimalID")
qpcr_data <- left_join(qpcr_data, stability, by="AnimalID")

qpcr_data$Sex <- as.factor(qpcr_data$Sex)
qpcr_data$Cage.x <- as.factor(qpcr_data$Cage.x)
qpcr_data$GrpSize <- as.factor(qpcr_data$GrpSize)
qpcr_data$dom_status_ds <- as.factor(qpcr_data$dom_status_ds)
qpcr_data$stab_status <- factor(qpcr_data$stab_status2, levels=c("Enforced stable", "Unstable", "Unenforced stable", "Unmeasurable"))
```

```{r}
Cyp11a1_m_f <- lm(rel_conc ~ Sex + ds + stab_status2 + ds*stab_status2, filter(qpcr_data, target.x=="Cyp11a1"))
Star_m_f <- lm(rel_conc ~ Sex + ds + stab_status2 + ds*stab_status2, filter(qpcr_data, target.x=="StAR"))
SF1_m_f <- lm(rel_conc ~ Sex + ds + stab_status2 + ds*stab_status2, filter(qpcr_data, target.x=="SF-1"))
DAX1_m_f <- lm(rel_conc ~ Sex + ds + stab_status2 + ds*stab_status2, filter(qpcr_data, target.x=="DAX1"))
Cyp11b1_m_f <- lm(rel_conc ~  Sex + ds + stab_status2 + ds*stab_status2, filter(qpcr_data, target.x=="Cyp11b1"))

print("Cyp11a1")
anova(Cyp11a1_m_f)
summary(Cyp11a1_m_f)
print("Star")
anova(Star_m_f)
r.squaredGLMM(Star_m_f)
print("SF-1")
anova(SF1_m_f)
print("DAX1")
anova(DAX1_m_f)
summary(Nrb01_m_f)
print("Cyp11b1")
anova(Cyp11b1_m_f)
summary(Cyp11b1_m_f)
```
```{r, correlation with age}
Cyp11a1 <- filter(qpcr_data, target.x=="Cyp11a1")
Cyp11b1 <- filter(qpcr_data, target.x=="Cyp11b1")
SF1 <- filter(qpcr_data, target.x=="SF-1")
DAX1 <- filter(qpcr_data, target.x=="DAX1")
StAR <- filter(qpcr_data, target.x=="StAR")


cor.test(Cyp11a1$rel_conc, Cyp11a1$AgeAtTest, method="spearman", exact=FALSE)
cor.test(Cyp11b1$rel_conc, Cyp11b1$AgeAtTest, method="spearman", exact=FALSE)
cor.test(DAX1$rel_conc, DAX1$AgeAtTest, method="spearman", exact=FALSE)
cor.test(SF1$rel_conc, SF1$AgeAtTest, method="spearman", exact=FALSE)
cor.test(StAR$rel_conc, StAR$AgeAtTest, method="spearman", exact=FALSE)
```


```{r, sex comparison}
genes <- qpcr_data %>% 
  filter(target.x != "GAPDH")

genes <- mutate(genes, Sex2=if_else(Sex=="F", "Female", "Male"))

sex_plot <- ggplot(genes, aes(Sex2, rel_conc, fill=Sex2))+
  stat_summary(geom="col", fun="mean", color="black")+
  stat_summary(geom="errorbar", fun.data="mean_se", width=0.5)+
  facet_grid(cols=vars(target.x))+
  scale_fill_npg()+
  scale_y_continuous(expand=c(0,0), limits=c(0,1.1))+
  scale_x_discrete(labels=c("Fale", "Male"))+
  xlab("") + ylab("")+
  labs(fill="Sex")+
  theme_cowplot()+
  theme(legend.position = "right", strip.text=element_text(face="italic"), axis.text.y=element_blank(), axis.text.x=element_text(angle=-90)) +
  panel_border()

sex_plot
```

```{r, age comparison}
age_plot <- ggplot(genes, aes(AgeAtTest, rel_conc))+
  geom_point(color="lightgrey", size=2)+
  geom_point(shape=1, color="black", size=2)+
  geom_smooth(method=lm, se=FALSE, color="black")+
  scale_y_continuous(expand=c(0,0), limits=c(0,1.1))+
  xlab("Age in days") + ylab("mRNA Expression")+
  facet_grid(cols=vars(target.x))+
  theme_cowplot()+
  theme(strip.text=element_text(face="italic"), axis.text.x=element_text(angle=-90), legend.position = "none")+
  panel_border()
  
age_plot
```


```{r, david's score comparison}

ds_plot <- ggplot(genes, aes(ds, rel_conc))+
  geom_point(aes(fill=ds), color="black", pch=21, size=2)+
  geom_smooth(method=lm, se=FALSE, color="black")+
  scale_x_continuous(limits=c(-3,3), breaks=c(-3,0,3))+
  scale_y_continuous(expand=c(0,0), limits=c(0,1.1))+
  scale_fill_gsea()+
  xlab("David's Score") + ylab("mRNA Expression")+
  labs(color="David's \nScore")+
  facet_grid(cols=vars(target.x))+
  theme_cowplot()+
  theme(strip.text=element_text(face="italic"), legend.position = "none")+
  panel_border()
  
ds_plot
```
```{r, stability category comparison}

genes$stab_status2 <- factor(genes$stab_status2, levels=c("Enforced stable", "Unstable", "Unenforced stable", "Unmeasurable"))

stbl_plot <- ggplot(genes, aes(stab_status2, rel_conc, fill=stab_status2))+
  stat_summary(geom="col", fun="mean", color="black")+
  stat_summary(geom="errorbar", fun.data="mean_se", width=0.5)+
  facet_grid(cols=vars(target.x))+
  scale_fill_npg()+
  scale_y_continuous(expand=c(0,0), limits=c(0,1.1))+
  xlab("") + ylab("")+
  guides(fill=guide_legend(title="Stability \nCategory", nrow=2, byrow=TRUE))+
  theme_cowplot()+
  theme(legend.position = "right", strip.text=element_text(face="italic"), axis.text.y=element_blank()) +
  panel_border()

stbl_plot
```
```{r}
intrxn_plot <- ggplot(genes, aes(ds, rel_conc))+
  geom_smooth(aes(color=stab_status2), method=lm, se=FALSE, size=1)+ 
  geom_point(aes(color=stab_status2), size=2)+
  geom_point(shape=1, size=2, color="black")+
  scale_x_continuous(limits=c(-3,3), breaks=c(-3,0,3))+
  scale_y_continuous(expand=c(0,0), limits=c(0,1.1))+
  scale_color_npg()+
  xlab("David's Score") + ylab("mRNA Expression")+
  facet_grid(cols=vars(target.x))+
  guides(color=guide_legend(title="Stability \nCategory", nrow=2, byrow=TRUE))+
  theme_cowplot()+
  theme(strip.text=element_text(face="italic"), legend.position = "none")+
  panel_border()

intrxn_plot
```


```{r}

qplot <- plot_grid(
  age_plot, sex_plot + theme(legend.position="none"), ds_plot, stbl_plot+theme(legend.position="none"), intrxn_plot, 
  ncol=2, labels="AUTO"
)

ggsave("/Users/justinvarholick/Documents/GitHub/AcomysDominance_2022_Data_Results/R generated figures/Fig6_qpcr_plots_rev_April.png", qplot, units="in", width = 9, height =9, device='png', dpi=300)

qplot

```
